Don't use LFN names.  Fix background uploader. Fix progress bar.  Bump version to 0.6

+++ src/deviation/Progress.java
@@ -1,7 +1,7 @@
 package deviation;
 
 public interface Progress {
-    public void update(Double val);
+    public void update(Integer val);
     public boolean cancelled();
 
 }
